pyproject.toml
[project]
name = "food-price-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6.0",
    "django-crudbuilder>=0.2.8",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
]
